+++ helpers/util/isExternalLink.ts
