+++ helpers/util/logTruthy.ts
