+++ helpers/voting/getVoteMetaData.ts
