+++ helpers/voting/makeUniqueKeyForVote.ts
