+++ helpers/web3/truncateEthAddress.ts
